SASTRUTS-123: version 1.0.1. supported S2 2.4.39. fixed issues found by findbugs.
git-svn-id: https://www.seasar.org/svn/sastruts/trunk/sa-struts-portlet@871 311f1561-e253-4dfb-a9fb-882cde7d7359

+++ src/main/java/org/seasar/struts/portlet/servlet/SAStrutsFilterChain.java
@@ -70,7 +70,7 @@ public class SAStrutsFilterChain implements FilterChain {
     public void doFilter(ServletRequest request, ServletResponse response)
             throws IOException, ServletException {
         // Call the next filter if there is one
-        int pos = ((Integer) filterPosition.get()).intValue();
+        int pos = filterPosition.get().intValue();
         if (pos < n) {
             Filter filter = filters[pos++];
             filterPosition.set(Integer.valueOf(pos));
